dE/dx distribution for electrons & protons passing through the straws
The straw hits allow us to reconstruct both total energy and dE/dx values. Bounds ("cuts") can be placed on these characteristic values to roughly distinguish between different particles and therefore these can be used to identify the particles that hit the straws. The electron hits constitute the desired "signal" for this experiment and the proton hits are considered to be undesired "background". Due to the overlap in the center between proton and electron distributions in both the above plots, it is not possible to choose a cut value that can segregate the electron and proton hits with 100% accuracy. Neither do we want to unwantedly ignore too many electron hits by keeping the cut value very low, nor do we want to incorrectly mistake protons for electrons by keeping the cut value too high. So the task at hand is to optimize the cut value so that the efficiency of signal detection is maximized while keeping the background rejection as efficient as possible.