Visual comparison of perfomance
These are the ROC curves generated by the manual optimization code. The red curve, corresponding to the use of a combination of E and dE/dx cuts, appears above the two two curves, hence suggesting that this strategy has indeed enhanced the proton rejection efficiencies for given electron acceptance efficiencies.