Optimization using TMVA
This code is based on a machine learning algorithm template used by the Flavor group, where changing some parts of the code template like the choice of training model (the group uses Multi Layer Perceptron generally), input variables sections, control statements and output file specifications should suffice to initiate the training. So it is a very useful setup for beginners in machine learning on TMVA. This code will automatically generate a plethora of plots that get displayed via a user-friendly GUI during runtime. Many of those plots were out the scope of my project requirement, but the combined ROC curve was essential and has been presented as an image in this repository. The curve does look quite similar to the one obtained "by hand".