ROC plot image from TMVA GUI
The ROC plot here for the optimized cut is the TMVA version of the one that we generated earlier without the use of machine learning concepts. They  strongly resemble each other, thus suggesting that our choice of training model (MLP) was reasonable.
The second plot from the GUI gives useful information about the cut efficiencies, but is a bit complicated and was analyzed further by my supervisor.