Similarity in dE/dx data for Muons and Electrons
This is the image that depicts the unexpected similarity between dE/dx data for muons and electrons in the model Mu2e experiment that my supervisor studied. This was spotted from one of the datasets that I was given. The group then planned to investigate this further using concepts related to the Bethe Bloch equation, but I did not have much more contribution in that. The corresponding code here is fairly straightforward because it just plots the data from a directory of a ROOT TTRee.